move config files to config folder. remove two_factor library from INSTALLED_APPS
