+++ config/celery.py
@@ -3,7 +3,7 @@ import os
 from celery import Celery
 from celery.schedules import crontab
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "apps.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "config.settings")
 app = Celery("interview-example")
 app.config_from_object("django.conf:settings", namespace="CELERY")
 app.autodiscover_tasks()
